Reveal I changed a lot of old stuff
