+++ app/db/__init__.py
@@ -1,1 +1,2 @@
-from .job_search import _job_search as searches
+from .jobsearch import _job_search as searches
+from .position import _position as positions
